play command is working
