+++ src/commands/music/queue.ts
@@ -1,23 +1,26 @@
 import { SlashCommandBuilder, CommandInteraction } from 'discord.js';
-import { serverQueue } from '../../lib/music-utils';
+import { distube } from '../../bot';
 
 export const data = new SlashCommandBuilder().setName('queue').setDescription('display all queued songs');
 
 export function execute(interaction: CommandInteraction): void {
-	if (serverQueue.getQueuedSongs().length === 0) {
+	const queue = distube.getQueue(interaction.guild?.id as string);
+
+	if (!queue) {
 		interaction.reply('There are no songs in the queue!').catch(console.error);
 
 		return;
 	}
 
-	const songListString: string[] = serverQueue
-		.getQueuedSongs()
-		.map(
-			(song, index) =>
-				`[${index + 1}]: ${song.info.videoDetails.title} requested by [${
-					song.requestingUser.username
-				}]`,
-		);
+	if (queue?.songs?.length === 0) {
+		interaction.reply('There are no songs in the queue!').catch(console.error);
+
+		return;
+	}
+
+	const songListString = queue.songs.map(
+		(song, index) => `[${index + 1}]: ${song.name} requested by [${song.user}]`,
+	);
 
 	const codeBlockFormattedString: string = '```ini\n' + `${songListString.join('\r\n')}` + '```';
